# Conflicts:
#	.gitignore

+++ cpp/src/command_classes/CentralScene.cpp
@@ -96,7 +96,7 @@ bool CentralScene::RequestState
 bool CentralScene::RequestValue
 (
 		uint32 const _requestFlags,
-		uint8 const _what,
+		uint16 const _what,
 		uint8 const _instance,
 		Driver::MsgQueue const _queue
 )
