+++ cpp/src/command_classes/Clock.cpp
@@ -90,7 +90,7 @@ bool Clock::RequestState
 bool Clock::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _dummy1,	// = 0 (not used)
+	uint16 const _dummy1,	// = 0 (not used)
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
