+++ cpp/src/command_classes/Language.cpp
@@ -76,7 +76,7 @@ bool Language::RequestState
 bool Language::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _dummy1,	// = 0 (not used)
+	uint16 const _dummy1,	// = 0 (not used)
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
