+++ cpp/src/command_classes/Protection.cpp
@@ -80,7 +80,7 @@ bool Protection::RequestState
 bool Protection::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _dummy1,	// = 0 (not used)
+	uint16 const _dummy1,	// = 0 (not used)
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
