+++ cpp/src/command_classes/SwitchMultilevel.cpp
@@ -130,7 +130,7 @@ bool SwitchMultilevel::RequestState
 bool SwitchMultilevel::RequestValue
 (
 	uint32 const _requestFlags,
-	uint8 const _index,
+	uint16 const _index,
 	uint8 const _instance,
 	Driver::MsgQueue const _queue
 )
